rename to TransformedStringType
