+++ packages/quicktype-core/src/Type/TransformedStringType.ts
